requirements.txt
Django==1.10.1
PyJWT==1.4.2
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
gunicorn==19.6.0
whitenoise==3.2.1
